zackplauche-backend/requirements.txt
asgiref==3.4.1
boto3==1.20.34
botocore==1.23.34
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
django-rest-framework==0.1.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
gunicorn==20.1.0
jmespath==0.10.0
Pillow==9.0.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
